Rename user routes
