+++ public/pages/user/register.js
@@ -1,8 +1,8 @@
 import { html } from "common-tags";
-import BasePage from "./basePage";
-import Form from "../components/form/form";
-import Input from "../components/form/elements/input";
-import SubmitButton from "../components/form/elements/submitButton";
+import BasePage from "../basePage";
+import Form from "../../components/form/form";
+import Input from "../../components/form/elements/input";
+import SubmitButton from "../../components/form/elements/submitButton";
 
 export default class Register extends BasePage {
     renderContent() {
